requirements.txt
asgiref==3.2.10
Django==3.0.8
django-filter==2.3.0
djangorestframework==3.11.0
Markdown==3.2.2
pytz==2020.1
sqlparse==0.3.1